requirements.txt
alembic==1.0.3
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
cffi==1.11.5
Click==7.0
colorama==0.4.0
coverage==4.5.2
cryptography==2.4.2
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
idna==2.7
intelecom==0.0.5
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
mysqlclient==1.3.13
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pydocstyle==3.0.0
pyflakes==2.0.0
PyJWT==1.6.4
pylama==7.6.6
pylint==2.1.1
pytest==4.0.0
pytest-cov==2.6.0
python-dateutil==2.7.5
python-editor==1.0.3
six==1.11.0
snowballstemmer==1.2.1
SQLAlchemy==1.2.14
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
